move getActionProceduresBase to own file
